pythonProject/requirements.txt
pandas
numpy==1.26.4
h5py==3.10.0
openai-whisper
networkx
sentence-transformers
transformers
bert-score
rouge-score
joblib
tqdm
xgboost
scenedetect[opencv]
scikit-learn